Events Added, Audio imported
Events were added to the game, there are only 2, tornado (fling objects randomly) and inverse gravity, variables can be set in the inspector, eg. how often it happens, the strength and time. If an event happens, there are words that show up that indicate that an event is happening. Events dont affect the player but rather the objects only.

I tried to do the audio of the game first, but I need to do more research(?), sounds dont happen, so I decided to just do the fun events first.